software/micromouse_analysis/requirements.txt
pyserial
more-itertools
matplotlib
numpy